refactor: merge AbstractInputValidator into TrivuleInput
Unify the input validation architecture by merging the abstract base class
directly into the concrete implementation. This simplifies the codebase by
removing an unnecessary abstraction layer.

Changes:
- Merged all members and methods from AbstractInputralidator into TrivuleInput
- Removed inheritance relationship and abstract base class
- Updated exports to remove abstract-input module
- Deleted src/core/abstract-input.ts

Benefits:
- Simpler architecture with single class instead of inheritance hierarchy
- Improved maintainability with all logic in one place
- Easier code navigation without jumping between files
- No breaking changes to public API

Closes #19
